+++ gladier_xpcs/tools/__init__.py
@@ -1,9 +1,6 @@
-
-
 from .corr import EigenCorr
 from .apply_qmap import ApplyQmap
 from .custom_pilot import CustomPilot
 from .plot import MakeCorrPlots
-from .manifest_reprocessing import ManifestReprocess
 
-__all__ = ['EigenCorr', 'ApplyQmap', 'CustomPilot', 'MakeCorrPlots', 'ManifestReprocess']
+__all__ = ['EigenCorr', 'ApplyQmap', 'CustomPilot', 'MakeCorrPlots']